Remove xDoubleLim
Summary:
This is no longer used.

This is a slight regression on Gameboy on the Pi 5. But it seems to be
some kind of alignment issue interfering with branch prediction. There
is no regression the Pi 3, and adding a single random mov to the setup
of the function seems to restore the 2-3% perf hit on the Pi 5.

Reviewed By: tmikov

Differential Revision: D65185513

fbshipit-source-id: 0fd0e372c291341ac8967203601f01397da2ee0e