Use a separate scope for params with expressions
Summary:
Parameters with expressions are supposed to create a new scope for
parameters only, so that functions in the parameters can't capture
variables in the function body.

This requires using a separate scope for parameters,
created prior to the scope for the function body.
Because the parameters may themselves create scopes (e.g. for function
expression names), the function body's scope is no longer at a constant
index in the `FunctionInfo`'s list of scopes, so we have to store the
index in `FunctionInfo`.

Having done that, conditionally create a new scope when there are
parameter expressions.

Reviewed By: tmikov

Differential Revision: D58532226

fbshipit-source-id: 36e2a814bc7608fa8c0235c320892bedb1713055