Add mechanism for deduplicating loop scopes
Summary:
When we create an inner scope for a loop, we need to ensure that the
same VariableScope is used every time the loop is generated.

For instance, if a loop appears in a finally block that is emitted
multiple times, we need to ensure that all those instances of the loop
use the same VariableScope. This ensures that inner functions in the
loop have a unique parent (since inner functions are uniqued).

This is done by associating a `VariableScope` with each loop AST node
in a map in the `FunctionContext`

Reviewed By: avp

Differential Revision: D59036599

fbshipit-source-id: 048b878fa2c792c1d185a320fa5575a0b1980e34