Handle for loops with body captures in the normal path
Summary:
Augment the normal for-loop path to create a new scope on each iteration
if it contains a function in its body. When we do this, create a set of
loop variables in both the body scope and the outer scope. After each
iteration, the variables are copied to the outer scope which holds them
between iterations, and for the update and test expressions (which we
know do not capture).

Reviewed By: tmikov

Differential Revision: D58274075

fbshipit-source-id: dfedd90acb1844456c236a873f4b035fc4d6909d