JIT: Add some TypeOfIs fast paths
Summary:
Split this off to make it easier to review.

The fast paths work for all the single-bit instructions and the
object|null case which is needed for `typeof x === "object"`.

The fast paths are also used for the inverses.

If we add the ability to merge `TypeOfIs` branches, we can add a more
general fast path - for now, the slow path is just a call.

Reviewed By: neildhar

Differential Revision: D65067396

fbshipit-source-id: 270149bf9099eceb0392c430fa8ddafc5dca4978