Reduce branching in multi-bit TypeOfIs
Summary:
Simplify the branching in TypeOfIs for multiple bits by initialising the
result register to the "match value". This way, we can jump directly to
the end if a match occurs, or fall through to the bit adjustment if no
match happens, allowing us to avoid the potential branch in the fall
through case.

Reviewed By: avp

Differential Revision: D65986419

fbshipit-source-id: a1475e77e7fc878ea126fe52e425c825c7ad42e1