Avoid MSVC iterator debugging assertion
Summary:
When compiling with MSVC STL headers and iterator debugging enabled
(a common case for debug builds), attempting to dereference
std::string::end() leads to a failing assertion.

To avoid this, we can just use the data pointer and string size to
fetch the begin and end addressed of the string.

Reviewed By: neildhar

Differential Revision: D66009100

fbshipit-source-id: cac2b796be4973db6811b08a2e696c8df7ec85d6