Make SHObjectFlags have a union form, so we can do bit operations to check for several flags at once.
Summary:
In a later diff (D65297624), it is useful to be able to check an
object for several flags at once.  To make this legal, we need to be
able to interpret the flags field as an integer.  This diff introduces
a union type that allows that.  This follows an idiom we use
elsewhere; for example, in PropertyDescriptor.h, for PropertyFlags.

Reviewed By: neildhar

Differential Revision: D65948788

fbshipit-source-id: 06fad3bf70a47697b6a63a0afcc148ab072f47d0