Revert "Disable DateTimeFormat::formatToParts for Apple platform (#1155)" (#1567)
Summary:
This reverts commit c5a633f6452902cd0b5c9392a1e4f51a33933284.

That commit removed the Intl.DateTimeFormat.formatToParts for Apple platforms from Hermes.
However, the community started depending on it, and we ended up manually reverting that commit in the release branch for the past 3 releases.

This is like reverting the commit in main, so we are proceeding with that approach to simplify the release process of React Native and Hermes to the OSS.

Pull Request resolved: https://github.com/facebook/hermes/pull/1567

Test Plan: CI

Reviewed By: blakef

Differential Revision: D66099176

Pulled By: cipolleschi

fbshipit-source-id: fe42bab7308bbd2294911be694f0bef8d1652566