Add a pass skeleton for optimizing known builtin calls
Summary:
This is a no-op diff.
It's split from the others to make it easier to review the fast paths.

We add a new pass `LowerBuiltinCallsOptimized` which will also emit
fast paths for known builtins (as of this diff there are none).

We add a `calleeNamesToOptimize` to `LowerBuiltinCallsContext`
which will contain the optimizable callees.

Reviewed By: fbmal7

Differential Revision: D66112033

fbshipit-source-id: 07ac543671f08a67ec7b617b0304c34e63ab2a5f