Add instructions to branch on builtin checks
Summary:
Add IR and Bytecode instructions to check and branch based on whether
a value matches the value of a builtin.

These aren't emitted yet, but they'll be used to emit fast paths,
for example `call` and `apply(arguments)`.

JIT unimplemented in this diff, added in D65348305.

Add the short versions of the bytecodes here because these are used
somewhat frequently.

Reviewed By: tmikov

Differential Revision: D65348307

fbshipit-source-id: 3f6a0dc11432b6b9222c1f85cbf6ae11fea38561