Align arithmetic slow paths with out-of-line cases
Summary:
Out-of-line interpreter cases just take the interpreter state as
parameters, which allows them to minimise the amount of work that has to
be done in the interpreter itself. Apply the same approach to arithmetic
operations in the interpreter, by having their slow paths take the
interpreter state and internally read/write registers.

This diff doesn't really move benchmarks, but together with the later
diffs, it significantly reduces the amount of code in the interpreter.

Reviewed By: tmikov

Differential Revision: D65736831

fbshipit-source-id: fd04e8bfb6c1a8cbba93ef43ff20fcd74d25be18