Move compilation StackExecutor to Runtime
Summary:
This allows us to use `StackExecutor` for both local and global `eval`.
Move the consideration of stack space into the VM, and the compiler no
longer uses it directly.

NOTE: As is, the diff deliberately keeps the changes in HBC.cpp small to
make it easier to pick into stable if necessary. We could refactor
things to avoid taking the `void *` argument in the worker functions.

Reviewed By: tmikov

Differential Revision: D66398356

fbshipit-source-id: a11539ecf028bd5c609ae8a880af8f0f1376edd2