Parse class static blocks
Summary:
Manual port of D65822124.

Class static blocks can't have return in them, so we have to check for
the Return param and error in the parser, instead of waiting for
SemanticValidator to error (where we'd have to track whether we are
currently in the top-level of a class static block).

We also have to forbid await expressions, which is easy to do by using
the forbidAwaitExpression_ flag in SemanticValidator.

The ESTree spec adds a StaticBlock node which is a BlockStatement.

NOTE: We have to make SH-specific changes for typed mode. We have to
allow return outside functions because we wrap in IIFE after the parse
step.

Reviewed By: fbmal7

Differential Revision: D66711283

fbshipit-source-id: 0b369d9cfd3f176f6c26d469c0960f038c6910cc