Port bytecode determinism test to static_h.
Summary:
I raised some questions about whether the compiler was deterministic.  Aakash noted that there was a test for the run-to-run determinism of at least the bytecode produced by hermesc in hermes, but that we weren't running it for static_h.

This diff ports that test over to static_h.  I changed the buck rule to be a sh_test, so we could run it as a standard test if we want.

I left the big file it uses, the Fb4a bundle, in the hermes directory.  I figure that we should only have one copy of big files; when we formally retire hermes someday, we can move those files.

allow-large-files

Reviewed By: lavenzg

Differential Revision: D66480990

fbshipit-source-id: a8eeed89a2d82205d9365d8f4d41abbd5f915418