Add a rudimentary implementation to CacheNewObject
Summary:
Add a first-attempt, fairly conservative set of rules to the
CacheNewObject optimization pass.

If a function's first uses of `this` are as the target objects
in property stores using literal keys function,
and they occur in a sequence in which earlier users dominate later ones,
and there's no `try` surrounding them, we can add
the CacheNewObject instruction. Otherwise, we bail.

The optimization only applies if there are MIN_PROPERTIES_FOR_CACHE
writes to literal `this` properties in a row.

Reviewed By: neildhar

Differential Revision: D40532426

fbshipit-source-id: 5c1ad5748f71528ecedcab3ea621c0d9daf5be62