Update UTF8 String emission and parsing in Synth Traces
Summary:
Updates the UTF8 String emission in Synth Traces. This includes the
following changes:

1. When emitting UTF-8 strings, any non-ascii character will be emitted
as the byte sequence of the UTF-8 encoding.

2. When decoding the UTF-8 string, the code point in JSONString is
expected to represent each byte of the UTF-8 String.

These changes will allow us to disable the logic that converts UTF8
(with surrogates) -> UTF16 -> UTF8 (without surrogates) in JSLexer, and
punt the decoding logic to SynthTrace code.

Reviewed By: neildhar

Differential Revision: D65763450

fbshipit-source-id: 749efb4c69a98aabc00bc01bded15d2103c82bd8