Add thread local caching to thread_stack_bounds
Summary:
To improve performance in situations where the runtime is repeatedly
called from different threads, add a second level of caching for the
stack bounds in a `thread_local` in OSCompat. Only if this cache fails
will we actually fall back to the platform-specific API calls.

The user-facing API doesn't change, we just split the
`thread_stack_bounds` code into a public function and an internal
`_impl` function.

Reviewed By: tmikov

Differential Revision: D66839832

fbshipit-source-id: 8505783721ae0d617792743d7413c84f51da75ae