Class fields with initializers
Summary:
Implement (static) class fields. This consists of creating a pair of
synthetic functions to do the work of writing these properties to the
correct object (either the instance or the class itself.) Additionally,
we must cache the value of the computed keys so that future invocations
of a class produce objects containing the same keys.

Reviewed By: neildhar

Differential Revision: D58444745

fbshipit-source-id: 83ce51d46856472bb6bc836b66d9f18739f48b9b