Lazy compilation of class methods
Summary:
Implement lazy compilation of class methods (excluding support for
constructors in this diff.) The core of this problem is being able to
restart compilation into `parseClassBodyImpl`. So, we need to make sure
that we are restarting at the correct source location. We do this by
modifying the location range of the IR function to point to the parent
MethodDefinitionNode.

Reviewed By: avp

Differential Revision: D65837264

fbshipit-source-id: 911295efc881fab5e06fe1ac601678c62827c525