Debugger eval in derived class constructor
Summary:
This requires persisting the `Variable *`s needed to restore
`LegacyClassContext`. The trickier part of the diff is making sure that
the IR for `this` is generated properly. This requires two changes.

First, we must treat the function containing the debugger eval
expression as an arrow function in IRGen. This is so that we correctly
load in the `CapturedState` when restarting IRGen. This also means that
`initCaptureStateInES5FunctionHelper` needs to broaden its criteria and
also initialize this captured state if the current function is a derived
class constructor.

Second, we must fix `nearestNonArrow` to skip over the debugger
function. This is because the logic for emitting a checked access on
`this` only fires if the nearest non-arrow is a derived class. But the
`FunctionInfo*` for the top-level debugger eval func does not have
`arrow` set. Rather than setting `arrow` to true, we can detect when a
`FunctionInfo*` is actually a debugger eval function. We do this by
creating a new flag thats true for top-level program nodes. We can then
deduce that if something is a program node but it's not the global
function, then it must be the debugger eval function.

Reviewed By: avp

Differential Revision: D66117737

fbshipit-source-id: 612720927501dc7843016456ea4f699692e77f85