Add a new Symbol IR Type
Summary:
We will introduce a new IR instruction for producing a symbol value, to
be used for the value of private names. In order to accurately express
its output type, we need a new type added in the IR type system. This
type will also be useful for writing asserts later in the stack- e.g.
there will be certain instructions which should only take in `Symbol`
values.

This is also useful for modeling the type of `ToPropertyKey`.

Reviewed By: avp

Differential Revision: D64867418

fbshipit-source-id: 67910462dfa5bc1f02a230e91b6a05a92568e581