Create derived class callables
Summary:
We need to be able to quickly verify that a given Callable is a derived
class constructor. Therefore, make a new CellKind for BC and native
derived class constructors.

Note that it's currently not possible to establish a single contiguous
arrangement of Callables to satisfy all the different ranges we want to
draw inside of the Callable CellKind space. But that's not necessarily
worse for perf currently, since `NativeJSFunction` now has to make check
two equality comparisons instead of two relational comparisons.

Reviewed By: avp

Differential Revision: D61358888

fbshipit-source-id: b12bc4a4b23881f8857cb01e20acb7fd1414c13e