Make StoreOwnPropertyInst throw on unsuccessful property write
Summary:
Make `StoreOwnPropertyInst` "strict". Practically speaking, this means
to throw if trying to write the property failed. It is correct to do
this because all of the existing usages of `StoreOwnPropertyInst` can
never throw. We only use it on object and array literals. Since we know
we just constructed these, they cannot have own non-writable properties
or be sealed.

This is a required change to make for classes, since we will be using
this instruction to define fields. There, it is possible to throw so we
must do so correctly.

Reviewed By: tmikov

Differential Revision: D66853055

fbshipit-source-id: 2a704f2f9c958b36f4bccaef097e803ac70b19fa