Always declare new scope for untyped class declarations
Summary:
Untyped classes require two variables to be created for a class
declaration. The inner one acts just like a ClassExpression variable.
Specifically, it should be treated as `const`, and be initialized. The
outer variable already acts as it should so there is no change required
there.

We don't want to interfere with the existing machinery for typed classes,
so now SemanticResolver needs to know if it's operating on typed JS.

Reviewed By: avp

Differential Revision: D65674255

fbshipit-source-id: 8b2533b5361a85cfe79dd20afe365c4b572fbeb2