Only create FuncInfo in typed classes for fields with value initializers
Summary:
The `ElementsInitFunction` means something different for typed vs legacy
classes. In typed classes, we know the shape of the object so we
preemptively emit an `AllocObjectLiteral` with the properties already
there, with default values. Therefore, when we encounter a field that
has no initializer, we don't need to create the `ElementsInitFunction`.

However, for legacy classes, we always need to be able to emit a
`DefineOwnPropertyInst` for any fields we encounter, including those
that do not have value initializers.

Reviewed By: avp

Differential Revision: D67115398

fbshipit-source-id: fc69db77b858480004e7530fc8e8105826cdcf3e