Add the requires cache.
Summary:
Continues the implementation of the CallRequire bytecode.

This diff adds the caching that actually makes this an optimization.
This turned out to just a little more subtle than I had originally
thought.  CallRequire is treated as a call.  On the first execution of
this bytecode for a given module id it must make the require call, and
then add the return result as the cache entry for the module id, to
make future CallRequire invocations for that module fast.  But it is
*not* the case that call instructions make the call, and return to the
intepreter case for the next instruction in the caller.  Instead, they
set up the stack and current-function information, and start executing
the callee.  Only on a ret instruction do we resume in the caller.  So
where do we populate the cache?

I solved this by generating a pair of bytecode instructions,
CallRequire and RequireCacheMiss.  The result of CallRequire is the
export value; RequireCacheMiss accepts this (and the module id as an
immediate), and populates the cache.  To make this work, the cache-hit
pass for CallRequire skips not just the CallRequire instruction, but
also the subsequent RequireCacheMiss instruction.

The cache is implemented in RuntimeModule -- a std::vector mapping
module ids (which should be dense non-negative integers) to the
"export value" of the module.  I had thought this was always a
non-null Object, but apparently it can be any value (and the Twilight
bundle has one case where the export is null).  We must take care to
scan this as a source of GC roots.

Reviewed By: neildhar

Differential Revision: D62135521

fbshipit-source-id: c3c5c2dfc6d8f448a3f6923851b066a73cdeeddc