Extend requires optimization to cover more cases.
Summary:
When running the require optimization on the Twilight bundle, it
printed a number of warnings, of the forms:
```
<unknown>:0: warning: 'require' escapes or is modified
```
and
```
<unknown>:0: warning: 'require' stored into local cannot be resolved
```

These turn out to have somewhat similar root causes.  The require
optimization works by starting at the module factory function argument
known to be the require function, and following the chain of its uses
(through other defs, for example).  If we reach a call, we know that
call is a require call.

In the cases that print the warnings, there's ambiguity.  In the first
case, a known-require value is an input to a phi.  In the second, we
store a require to a stack location.  The existing code handled that case
when there was only a single store, but in the Twilight bundle
examples there were multiple stores.  However, they all stored
require.

This diff extends the optimization to cover these cases, using similar
techniques.  A phi instruction has a fixed set of arguments (actually
argument pairs -- values and the BB from which the value came).  We
keep a map from phi instructions to the set of their argument indices
known to receive a require.  As a known-require value reaches a phi
use, we update that set for the phi instruction.  If this makes all
the phi arguments require, then we consider the phi to be
known-require, and consider its uses.

Somewhat similarly, for the stack location case, we keep a map from
AllocStackInst* to a set of StoreStack instruction that store the
require value.  We update this as we find StoreStacks that store
require.  When we've otherwise exhausted the worklist, we check
whether this has identified any AllocStackInsts *all* of whose stores
store require.  If so, we consider the loads from that stack location
to be require uses.

(As an editorial comment: this sort of thing would be considerably
simpler, more complete, and re-usable if we did this with a
general-purpose flow analysis framework.  Though admittedly perhaps
somewhat slower...though I'm not sure that really matters.)

Reviewed By: avp

Differential Revision: D62645781

fbshipit-source-id: 580f12f19c5bc271788db6eac9b5b57192d6f27d