ignore blackboxed ranges when debugging
Summary:
Implement `Debugger.setBlackboxedRanges` that would save a list of all the ranges that stepping into would trigger a subsequent step.
This is also used in filtering out frames in stack traces coming from these ranges.

Notice that this request (`Debugger.setBlackboxedRanges`) is not the same as `Debugger.setBlackboxPatterns` that can be set up by the user in settings, that would be implemented in next diffs separately sharing some of the code with these methods.

Reviewed By: dannysu

Differential Revision: D62762752

fbshipit-source-id: 2fb720679dd4eec9e5d0c98dadbf3cc9deac3003