add scriptLanguage="JavaScript" to Debugger.scriptParsed
Summary:
The option to "ignore scripts" from the context menu only works for JS files.

The type of file is set when the debugger raises the CDP command [Debugger.scriptParsed](https://chromedevtools.github.io/devtools-protocol/tot/Debugger/#event-scriptParsed) using the `scriptLanguage` field.

That field was not set before.

Now it's sent as "JavaScript". (the other possible option is WebAssembly" which we don't support). See:
* https://github.com/ChromeDevTools/devtools-frontend/blob/7a7a8a4ba54c30d7e00ec5aee105910c57466a06/front_end/panels/sources/SourcesPanel.ts#L968
* https://source.chromium.org/chromium/chromium/src/+/main:v8/src/inspector/v8-debugger-agent-impl.cc;l=1875-1882

Reviewed By: dannysu

Differential Revision: D66101955

fbshipit-source-id: 501b44782f6aed145fd145ee957ef0a9bcbcdd91