CSE for requires calls.
Summary:
Previous work in this stack on require calls has focused on making
them faster, by caching the results for a particular module index.
This cache is global to an execution.  This diff attempts to identify
redundant require calls statically within a function: essentially,
making "require(n)" subject to CSE, by noting that require has special
semantics that allow this.  In particular, require(n) is idempotent
(wrt subsequent calls for the same value of n): the first call may
have side effects, but the second and subsequent calls do not.  We
make a special case for [metro-require] calls in CSEValue::canHandle
to reflect this assumption.

I add a statistic that breaks out the number of "require CSEs".

I also considered doing a specialized optimization for require calls
in Instruction::isIdenticalTo: instructions only get the
[metro-require] attribute if we've proved that their first argument is
the metroRequire function.  So we'd assume the first argument to the
all such calls are semantically the same in isIdentical (and in the
hash function).

This yielded a small number of additional require CSEs: on the order
of 25 to 30.  I give a range here because, try as I might, doing this
made the results (in terms of # of CSE) non-deterministic.  I was
unable to figure out why -- I took care to make sure the hash function
gave the same result whenever isIdentical would return true.  I
eventually decided that figuring out the mystery was not worth it,
given the small extra benefit.

Design question: in canHandle, we make a special case for
metro-require.  This seems perhaps less general than it might.  In
that function, we would all a call to be handled if it was labeled
[pure].  The metro-require case is "as if pure"; for a given module
Id, it has side effects, but only on the first call globally.  So
maybe we should have an [idempotent] attribute, in case we identify
other functions with this property, and generalize canHandle in this
way (pure || idempotent)?  (The special assumption of equivalent first
arguments is, however, specific to metro-require calls.)

Reviewed By: neildhar

Differential Revision: D64773199

fbshipit-source-id: d78538ac2cc12f10e6cfbaa4cc0879a826d09365