Fix handle error in requires opt.
Summary:
fbmal7 noticed that my previous diff stack ending in D62135521
caused at least one non-deterministic test failure, with handlesan.
This diff fixes the cause of that.

I left a raw HermesValue exposed across an allocating operation, and
we missed it.  Now we store it in a Locals struct, and retrieve it
from there after.

This is in the slow path of a requires cache miss, so the extra cost
is a negligable concern.

Reviewed By: neildhar

Differential Revision: D67166304

fbshipit-source-id: c21413186bf828a9c44211c015ba383887fa94a2