Align ConsecutiveStringStorage to 2-bytes before appending
Summary:
The `ConsecutiveStringStorage::appendStorage` function needs to align
the existing `storage_` before appending the `rhs`, because the `rhs`
has been constructed assuming that the storage begins at 0, which is
2-byte aligned. If we're appending onto a storage of an odd size, any
UTF-16 strings in the `rhs` become unaligned, resulting in UBSAN
crashes.

This function is only used to append onto a non-empty
`ConsecutiveStringStorage` in lazy compilation, so add a test that
demonstrates that this now works in lazy mode.

Reviewed By: tmikov

Differential Revision: D67368118

fbshipit-source-id: db3bcf20ce434c4b53a77a927fd0dedf48306f03