Treat `CreateThis` as leaking the function
Summary:
The resulting value from a construct call (via the `new` operator)
should be considered as leaking the function. This is because the
constructed `this` can access the function via its parent's
`.constructor` property.

This may pessimize inlining, if there was previously only a single
callsite. It also hurts type inference, since the types of the parameters
can no longer be inferred.

Reviewed By: avp

Differential Revision: D67207850

fbshipit-source-id: 35dda62f6b01d090b3fc16a7af66c7809654efe6