Use symbols for lookup instead of converting in for-in loops
Summary:
To speed up lookups in `GetNextPName`, keep the symbols around in the
`GetPNameList` and convert to strings later. This allows the existence
check to proceed faster than it used to, though it does still have to do
the `HiddenClass::findProperty` check.

`getForInPropertyNames` isn't just used in the Interpreter.
This diff audits the remaining callsites and handles Symbols there as
well.

Add a microbenchmark.
This diff makes it go from 1750ms to 1520ms.

Next Steps:
* Find a way to remove the `findProperty` check entirely.
* Optimize the GetByVal calls in the loop itself.

Reviewed By: tmikov

Differential Revision: D67054377

fbshipit-source-id: 406971971a406882a0e53fbe97da6fa6f9dc2dda