Add transform for match statements
Summary:
Original Author: gkz@meta.com
Original Git: 6e8e0db33510736230f3b7aa10d2efdc2258ac88
Original Reviewed By: SamChou19815
Original Revision: D66852555

This builds on the previous diff implementing the transform for match expressions. See that for more details.

Output:
We replace the match statement with a labelled block containing a series of if statements. We use the label to exit this block when we have a successful match. Otherwise, we go to the next test.

Reviewed By: fbmal7

Differential Revision: D67867910

fbshipit-source-id: 466d712e137ab624f897291da20c447d7874fcea