Pipe information through compiler for simplified resolver generation for property lookup resolvers
Summary:
Original Author: evanyeung@meta.com
Original Git: 28ed217412f0336b67b083b32f5544fc69c48450
Original Reviewed By: captbaritone
Original Revision: D66714731

@public
This diff creates the foundation of a new quality of life feature that allows for quickly defining resolvers that just do a property lookup on the base model. The changes in this diff are just piping the information of whether or not the resolver is a regular function or a new property lookup resolver through the compiler. It starts in the schema generation module and must carry the information all the way through to the printer. A stacked diff will actually implement the logic to find and add the resolver information.

Reviewed By: fbmal7

Differential Revision: D67867909

fbshipit-source-id: c2bb53dde62d793254c4b05a7bce8d1e5e7c4dd9