Improve match transform logic for determining if an 'in' check is needed for an object property's pattern
Summary:
The most basic example is:

```
const e = match (x) {
  {foo: undefined}: 0,
}
```

Previously we only outputted a check that `x.foo === undefined`, but if `foo` doesn't exist, this would still pass (e.g. input `{}`). To fix this, add an `in` check in that scenario. Also properly handle As and Or Patterns. Note that since an Identifier or Member Pattern may resolve to `undefined`, we add the `in` check for all those cases.

Reviewed By: alexmckenley

Differential Revision: D67777384

fbshipit-source-id: 4d348f6ea306c6aee741f1d5d939fb9e17227a63