Split the read and write property caches.
Summary:
The remainder of this stack was accepted, but Tzvetan pointed out that
it was adding a field to the PropertyCacheEntry that was used only for
GetById, not by PutById.  This was both at least a little confusing,
and wasted space.

This diff splits the property caches more completely into two caches,
so the entries can be different types (There had been a single array,
with an index recording the split point separating the Get and Put
entries.)  It creates the two types, though for this diff leaves them
as having the same fields.

Should not affect behavior as all.  Should also not affect
performance: before, the put cache was found by adding the
`writePropCacheOffset_` to the start of the
`TrailingObjects<PropertyCacheEntry>` of a `CodeBlock`.  After this
diff, there are two `TrailingObjects` arrays; I assume that it will
compile down to about the same code -- to get the second
`TrailingObjects` array, for the `WritePropertyCache entries, it would
call `numTrailingObjects` for the `ReadPropertyCache` array.  This
should come down, again, to adding a single in to the base of the
first `TrailingObjects` array.  This seems borne out by measurements.

Reviewed By: neildhar

Differential Revision: D67603981

fbshipit-source-id: e58c350ab6867639fee5d0c9865973bbdfc56c6c